refactor(parallelsearch): isolate transient PEP and merged-mode behind dedicated seams
- Extract TransientPepAnalysisEngine : PepAnalysisEngine under
  EngineLayer/ParallelSearch so the generic PEP engine keeps only its standard
  cross-validated workflow; the train-once / apply-to-transient model reuse and
  the background PEP->PEP_QValue curve move to the derived class. Only the
  _randomSeed seam moves private->protected (mirrors TransientProteinParsimony/
  ScoringAndFdr engines).
- Replace the MM_PARALLELSEARCH_MERGED environment variable with a real
  ParallelSearchParameters.UseMergedTransientLibrary flag so merged-mode is on
  the parameter/TOML/CLI surface and reproducible instead of process state.
- Per-database output now counts a match confident by EITHER its PEP_QValue or
  the score-based q-value when PEP is active, so the few reasonable edge-case
  hits a single metric flags are still written for the user to investigate.
  The membership filter (FilterToTransientDatabaseOnly) stays separate from the
  shared row-level confidence filter.

Addresses nbollis review on nbollis/MetaMorpheus#10.
